bus_test and AB_register_test work.
The problems I was having were wiring errors; first,
I forgot to switch the Arduino out signal from pin 53
to the '595s, so I added a message reminding me to
do that.  Second, I had the output enables of A and B
swapped.

Also added a boatload of diagnositcs to AB_register_test.